requirements.txt
langchain
python-dotenv
unstructured
faiss-cpu
libmagic
python-magic
python-magic-bin
sentence-transformers